Add low memory mode for supported utilities

+++ src/modules/PowerOCR/PowerOCRModuleInterface/dllmain.cpp
@@ -5,6 +5,7 @@
 #include "trace.h"
 #include "Generated Files/resource.h"
 #include <common/logger/logger.h>
+#include <common/SettingsAPI/settings_helpers.h>
 #include <common/SettingsAPI/settings_objects.h>
 #include <common/utils/resources.h>
 
@@ -60,7 +61,7 @@ private:
     //contains the non localized key of the powertoy
     std::wstring app_key;
 
-    HANDLE m_hProcess;
+    HANDLE m_hProcess = nullptr;
 
     // Time to wait for process to close after sending WM_CLOSE signal
     static const int MAX_WAIT_MILLISEC = 10000;
@@ -109,16 +110,43 @@ private:
 
     bool is_process_running()
     {
+        if (m_hProcess == nullptr)
+        {
+            return false;
+        }
+
         return WaitForSingleObject(m_hProcess, 0) == WAIT_TIMEOUT;
     }
 
-    void launch_process()
+    void launch_process(bool activate_on_startup = false, bool exit_after_close = false)
     {
-        Logger::trace(L"Starting TextExtractor process");
+        if (is_process_running())
+        {
+            Logger::trace("TextExtractor process is already running. Skipping launch.");
+            Logger::flush();
+            return;
+        }
+
+        if (m_hProcess != nullptr)
+        {
+            CloseHandle(m_hProcess);
+            m_hProcess = nullptr;
+        }
+
+        Logger::info(L"Starting TextExtractor process. activate_on_startup={}, exit_after_close={}", activate_on_startup, exit_after_close);
+        Logger::flush();
         unsigned long powertoys_pid = GetCurrentProcessId();
 
-        std::wstring executable_args = L"";
-        executable_args.append(std::to_wstring(powertoys_pid));
+        std::wstring executable_args = std::to_wstring(powertoys_pid);
+        if (activate_on_startup)
+        {
+            executable_args.append(L" --activate");
+        }
+
+        if (exit_after_close)
+        {
+            executable_args.append(L" --exit-after-close");
+        }
 
         SHELLEXECUTEINFOW sei{ sizeof(sei) };
         sei.fMask = { SEE_MASK_NOCLOSEPROCESS | SEE_MASK_FLAG_NO_UI };
@@ -127,14 +155,15 @@ private:
         sei.lpParameters = executable_args.data();
         if (ShellExecuteExW(&sei))
         {
-            Logger::trace("Successfully started the TextExtractor process");
+            m_hProcess = sei.hProcess;
+            Logger::info("Successfully started the TextExtractor process. pid={}", GetProcessId(m_hProcess));
+            Logger::flush();
         }
         else
         {
             Logger::error(L"TextExtractor failed to start. {}", get_last_error_or_default(GetLastError()));
+            Logger::flush();
         }
-
-        m_hProcess = sei.hProcess;
     }
 
     // Load the settings file.
@@ -241,7 +270,16 @@ public:
     {
         Logger::trace("TextExtractor::enable()");
         ResetEvent(m_hInvokeEvent);
-        launch_process();
+        ResetEvent(m_hTerminateEvent);
+        PTSettingsHelper::refresh_low_memory_settings_cache();
+        const bool low_memory_mode = PTSettingsHelper::is_low_memory_mode_enabled(get_key());
+        Logger::info("TextExtractor::enable(): low_memory_mode={}", low_memory_mode);
+        Logger::flush();
+        if (!low_memory_mode)
+        {
+            launch_process();
+        }
+
         m_enabled = true;
         Trace::EnablePowerOCR(true);
     };
@@ -252,9 +290,17 @@ public:
         if (m_enabled)
         {
             ResetEvent(m_hInvokeEvent);
-            SetEvent(m_hTerminateEvent);
-            WaitForSingleObject(m_hProcess, 1500);
-            TerminateProcess(m_hProcess, 1);
+            if (m_hProcess != nullptr)
+            {
+                SetEvent(m_hTerminateEvent);
+                if (WaitForSingleObject(m_hProcess, 1500) == WAIT_TIMEOUT)
+                {
+                    TerminateProcess(m_hProcess, 1);
+                }
+
+                CloseHandle(m_hProcess);
+                m_hProcess = nullptr;
+            }
         }
 
         m_enabled = false;
@@ -268,10 +314,22 @@ public:
             Logger::trace(L"TextExtractor hotkey pressed");
             if (!is_process_running())
             {
-                launch_process();
+                const bool low_memory_mode = PTSettingsHelper::is_low_memory_mode_enabled(get_key());
+                Logger::trace("TextExtractor hotkey cold start. low_memory_mode={}", low_memory_mode);
+                launch_process(true, low_memory_mode);
+                return true;
             }
 
-            SetEvent(m_hInvokeEvent);
+            if (!SetEvent(m_hInvokeEvent))
+            {
+                Logger::error(L"TextExtractor failed to signal invoke event. {}", get_last_error_or_default(GetLastError()));
+                Logger::flush();
+            }
+            else
+            {
+                Logger::trace("TextExtractor invoke event signaled.");
+            }
+
             return true;
         }
